refactor: rename components to Components
