+++ rs-react-app/src/App/App.tsx
@@ -1,7 +1,7 @@
 import './App.css';
-import { Footer } from '../components/Footer/Footer';
-import { Header } from '../components/Header/Header';
-import { Main } from '../components/Main/Main';
+import { Footer } from '../components/Footer-tmp/Footer';
+import { Header } from '../components/Header-tmp/Header';
+import { Main } from '../components/Main-tmp/Main';
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
