+++ rs-react-app/src/components/Footer/Footer.tsx
