+++ rs-react-app/src/components/ItemData/ItemData.tsx
