+++ rs-react-app/src/components/ThrowErrorButton/ThrowErrorButton.tsx
