+++ rs-react-app/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import { App } from './App/App.tsx';
-import { ErrorBoundary } from './components/ErrorBoundary/ErrorBoundary.tsx';
+import { ErrorBoundary } from './components/ErrorBoundary-tmp/ErrorBoundary.tsx';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { store } from './store/store.ts';
