+++ rs-react-app/src/pages/About/About.tsx
